Separated Authorization logic to ms-common-auth
Introduced an audience claim and validation.

Removed "audience" setting.  The audience is the one that is associated
when the token gets put into the cache.

The "get" operation removes the need for the Client ID to be passed into
the handler, since there is a check to make sure the client ID is
allowed to use that grant.

Removed oauth objects from common to keep the user facing code as little
as possible.