OIDC refinements
Converted state from the client to be stored temporarily in the cache and hidden from the OpenID IP.

This also adds the notion of additional claims so other information can
be passed down.

See #23, which will introduce a sample so the client and issuer data can
be parsed in a more application specific way.